+++ docs/data/material/getting-started/templates/shared-theme/AppTheme.tsx
@@ -1,0 +1,57 @@
+/**
+ * TODO: need to update the `updateTemplatesTheme.ts` script to include the components that are used in the template.
+ * Generated by a script.
+ */
+import * as React from 'react';
+import { ThemeProvider, createTheme } from '@mui/material/styles';
+import type { ThemeOptions } from '@mui/material/styles';
+import { inputsCustomizations } from './customizations/inputs';
+import { dataDisplayCustomizations } from './customizations/dataDisplay';
+import { feedbackCustomizations } from './customizations/feedback';
+import { navigationCustomizations } from './customizations/navigation';
+import { surfacesCustomizations } from './customizations/surfaces';
+import {
+  colorSchemes,
+  cssVariables,
+  typography,
+  shadows,
+  shape,
+} from './themePrimitives';
+
+export default function AppTheme({
+  children,
+  disableCustomTheme,
+  themeComponents,
+}: {
+  children: React.ReactNode;
+  disableCustomTheme?: boolean;
+  themeComponents?: ThemeOptions['components'];
+}) {
+  const theme = React.useMemo(() => {
+    return disableCustomTheme
+      ? {}
+      : createTheme({
+          colorSchemes,
+          cssVariables,
+          typography,
+          shadows,
+          shape,
+          components: {
+            ...inputsCustomizations,
+            ...dataDisplayCustomizations,
+            ...feedbackCustomizations,
+            ...navigationCustomizations,
+            ...surfacesCustomizations,
+            ...themeComponents,
+          },
+        });
+  }, [disableCustomTheme, themeComponents]);
+  if (disableCustomTheme) {
+    return children;
+  }
+  return (
+    <ThemeProvider theme={theme} disableTransitionOnChange>
+      {children}
+    </ThemeProvider>
+  );
+}
