--- docs/data/material/getting-started/templates/sign-in/theme/customizations/index.ts
@@ -1,5 +1,0 @@
-export { inputsCustomizations } from './inputs';
-export { dataDisplayCustomizations } from './dataDisplay';
-export { feedbackCustomizations } from './feedback';
-export { navigationCustomizations } from './navigation';
-export { surfacesCustomizations } from './surfaces';
