--- docs/data/material/getting-started/templates/sign-in/theme/getSignInTheme.js
@@ -1,21 +1,0 @@
-import { getDesignTokens } from './themePrimitives';
-import {
-  inputsCustomizations,
-  dataDisplayCustomizations,
-  feedbackCustomizations,
-  navigationCustomizations,
-  surfacesCustomizations,
-} from './customizations';
-
-export default function getSignInTheme(mode) {
-  return {
-    ...getDesignTokens(mode),
-    components: {
-      ...inputsCustomizations,
-      ...dataDisplayCustomizations,
-      ...feedbackCustomizations,
-      ...navigationCustomizations,
-      ...surfacesCustomizations,
-    },
-  };
-}
